+++ tests/test_atoms.py
@@ -9,8 +9,7 @@ def test_functions():
 
     assert atoms.get_maximal_valance(atom_label='C') == 4
     assert atoms.get_maximal_valance(atom_label='Aa') == 6
-    assert 11.9 < atoms.get_atomic_weight(atom_label='C') < 12.1
-    assert atoms.get_atomic_weight(atom_label='Aa') == 70
+    assert 11.9 < Atom('C').weight.to('amu') < 12.1
     assert 0.9 < atoms.get_vdw_radius(atom_label='H') < 1.2
     assert 2 < atoms.get_vdw_radius(atom_label='Aa') < 3
 
@@ -22,6 +21,8 @@ def test_functions():
 def test_atoms():
 
     empty_atoms = Atoms()
+    assert 'atoms' in repr(empty_atoms).lower()
+    assert not empty_atoms.are_linear()
 
     # Undefined COM with no atoms
     with pytest.raises(ValueError):
@@ -29,8 +30,18 @@ def test_atoms():
 
     h_atoms = Atoms([Atom('H'), Atom('H', x=1.0)])
     assert isinstance(h_atoms.com, Coordinate)
-
     assert np.allclose(h_atoms.com, np.array([0.5, 0.0, 0.0]))
+
+    assert h_atoms.vector(0, 1) == np.array([1.0, 0.0, 0.0])
+
+    # Moment of inertia
+    assert np.sum(np.diag(h_atoms.moi)) > 0.0
+    assert 1.9 < np.sum(np.diag(h_atoms.moi)) < 2.1
+
+    h_atoms_far = Atoms([Atom('H'), Atom('H', x=10.0)])
+    assert np.sum(h_atoms_far.moi) > np.sum(h_atoms.moi)
+
+    assert np.isclose(np.linalg.norm(h_atoms_far.nvector(0, 1)), 1.0)
 
     # COM is weighted by mass, so the x-coordinate
     ch_atoms = Atoms([Atom('H'), Atom('C', x=1.0)])
@@ -45,6 +56,7 @@ def test_atom_collection_base():
     h2 = atoms.AtomCollection()
     assert h2.n_atoms == 0
     assert h2.coordinates is None
+    assert h2.moi is None and h2.com is None
 
     # Cannot set coordinates without atoms
     with pytest.raises(ValueError):
@@ -134,12 +146,16 @@ def test_atom_collection_dihedral():
                       100.8,
                       atol=1.0)
 
-    # Undefined dihedral with a zero vector between teo atoms
+    # Undefined dihedral with a zero vector between two atoms
     with pytest.raises(ValueError):
         h2o2.atoms[0].coord = np.zeros(3)
         h2o2.atoms[1].coord = np.zeros(3)
 
         _ = h2o2.dihedral(2, 0, 1, 3)
+
+    # and a dihedral with atoms not present in the molecule
+    with pytest.raises(ValueError):
+        _ = h2o2.dihedral(2, 0, 1, 10)
 
 
 def test_atom():
@@ -183,9 +199,15 @@ def test_atom():
     assert dummy.atomic_number == 0
     assert dummy.period == 0
     assert dummy.group == 0
+    assert dummy.mass == 0.0
 
     fe = Atom(atomic_symbol='Fe')
     assert fe.tm_row == 1
+
+    # Should have a mass, even if it's estimated for all elements
+    for element in atoms.elements:
+        atom = Atom(element)
+        assert atom.weight is not None
 
 
 def test_atom_coord_setting():